// webapp/src/lib/rpc.gen.ts
export enum Direction {
  left = 'left',
  right = 'right',
  up = 'up',
  down = 'down'
}

export enum ItemType {
  bite = 'bite'
}

export interface Update {
  diffs: Array<Diff>
  state: State
}

export interface Diff {
  x: number
  y: number
  color: string
  add: boolean
}

export interface State {
  width: number
  height: number
  snakes: {[key: number]: Snake}
  items: {[key: number]: Item}
}

export interface Snake {
  id: number
  name: string
  color: string
  body: Array<Coordinate>
  direction: Direction
  nextDirections: Array<Direction>
  length: number
  bornAt: string
  diedAt: string
}

export interface Item {
  id: number
  color: string
  type: ItemType
  coordinate: Coordinate
}

export interface Event {
}

export interface Coordinate {
  x: number
  y: number
}

export interface SnakeGame {
  joinGame(options: WebrpcStreamOptions<JoinGameReturn>): Promise<void>
  createSnake(args: CreateSnakeArgs, headers?: object, signal?: AbortSignal): Promise<CreateSnakeReturn>
  turnSnake(args: TurnSnakeArgs, headers?: object, signal?: AbortSignal): Promise<TurnSnakeReturn>
}

export interface JoinGameArgs {
}

export interface JoinGameReturn {
  update: Update
  event: Event  
}
export interface CreateSnakeArgs {
  username: string
}

export interface CreateSnakeReturn {
  snakeId: number  
}
export interface TurnSnakeArgs {
  snakeId: number
  direction: Direction
}

export interface TurnSnakeReturn {  
}


  
//
// Client
//
export class SnakeGame implements SnakeGame {
  protected hostname: string
  protected fetch: Fetch
  protected path = '/rpc/SnakeGame/'

  constructor(hostname: string, fetch: Fetch) {
    this.hostname = hostname
    this.fetch = (input: RequestInfo, init?: RequestInit) => fetch(input, init)
  }

  private url(name: string): string {
    return this.hostname + this.path + name
  }
  
  joinGame = (options: WebrpcStreamOptions<JoinGameReturn>): Promise<void> => {
    const _fetch = () => this.fetch(this.url('JoinGame'),createHTTPRequest({}, options.headers, options.signal)
      ).then(async (res) => {
        await sseResponse(res, options, _fetch);
    }, (error) => {
      options.onError(error, _fetch);
    });
    return _fetch();
  }
  
  createSnake = (args: CreateSnakeArgs, headers?: object, signal?: AbortSignal): Promise<CreateSnakeReturn> => {
    return this.fetch(
      this.url('CreateSnake'), createHTTPRequest(args, headers, signal)).then((res) => {
        return buildResponse(res).then(_data => {
          return {
            snakeId: <number>(_data.snakeId),
          }
          })}, (error) => {throw WebrpcRequestFailedError.new({ cause: `fetch(): ${error.message || ''}` })}
      )
  }
  
  
  turnSnake = (args: TurnSnakeArgs, headers?: object, signal?: AbortSignal): Promise<TurnSnakeReturn> => {
    return this.fetch(
      this.url('TurnSnake'), createHTTPRequest(args, headers, signal)).then((res) => {
        return buildResponse(res).then(_data => {
          return {}
          })}, (error) => {throw WebrpcRequestFailedError.new({ cause: `fetch(): ${error.message || ''}` })}
      )
  }
  
  
}
// ...
export class WebrpcError extends Error {
  name: string
  code: number
  message: string
  status: number
  cause?: string

  /** @deprecated Use message instead of msg. Deprecated in webrpc v0.11.0. */
  msg: string

  constructor(name: string, code: number, message: string, status: number, cause?: string) {
    super(message)
    this.name = name || 'WebrpcError'
    this.code = typeof code === 'number' ? code : 0
    this.message = message || `endpoint error ${this.code}`
    this.msg = this.message
    this.status = typeof status === 'number' ? status : 0
    this.cause = cause
    Object.setPrototypeOf(this, WebrpcError.prototype)
  }

  static new(payload: any): WebrpcError {
    return new this(payload.error, payload.code, payload.message || payload.msg, payload.status, payload.cause)
  }
}

// Webrpc errors

export class WebrpcEndpointError extends WebrpcError {
  constructor(
    name: string = 'WebrpcEndpoint',
    code: number = 0,
    message: string = 'endpoint error',
    status: number = 0,
    cause?: string
  ) {
    super(name, code, message, status, cause)
    Object.setPrototypeOf(this, WebrpcEndpointError.prototype)
  }
}

export class WebrpcRequestFailedError extends WebrpcError {
  constructor(
    name: string = 'WebrpcRequestFailed',
    code: number = -1,
    message: string = 'request failed',
    status: number = 0,
    cause?: string
  ) {
    super(name, code, message, status, cause)
    Object.setPrototypeOf(this, WebrpcRequestFailedError.prototype)
  }
}
// ...
export type Fetch = (input: RequestInfo, init?: RequestInit) => Promise<Response>

export interface WebrpcStreamOptions<T> extends WebrpcOptions {
  onMessage: (message: T) => void;
  onError: (error: WebrpcError, reconnect: () => void) => void;
  onOpen?: () => void;
  onClose?: () => void;
}
export interface WebrpcOptions {
  headers?: HeadersInit;
  signal?: AbortSignal;
}